# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chameleon Golf API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Course {
  id: number;
  name: string;
  location?: string;
  holes: number;
  par: number;
  createdAt: string;
}

export interface CreateCourse {
  name: string;
  location?: string;
  holes: number;
  par: number;
}

export interface Round {
  id: number;
  courseId: number;
  courseName?: string;
  playerName: string;
  date: string;
  totalScore?: number;
  totalPar?: number;
  createdAt: string;
}

export interface CreateRound {
  courseId: number;
  playerName: string;
  date: string;
}

export interface Score {
  id: number;
  roundId: number;
  holeNumber: number;
  par: number;
  strokes: number;
  notes?: string;
}

export interface RoundWithScores {
  id: number;
  courseId: number;
  courseName?: string;
  playerName: string;
  date: string;
  totalScore?: number;
  totalPar?: number;
  scores: Score[];
  createdAt: string;
}

export interface CreateScore {
  holeNumber: number;
  par: number;
  strokes: number;
  notes?: string;
}

export type SaveScoresBody = {
  scores: CreateScore[];
};