# lib/api-zod/src/generated/types/createCourse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chameleon Golf API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCourse {
  name: string;
  location?: string;
  holes: number;
  par: number;
}