# lib/api-zod/src/generated/types/createRound.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chameleon Golf API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRound {
  courseId: number;
  playerName: string;
  date: Date;
}