# lib/api-zod/src/generated/types/roundWithScores.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chameleon Golf API
 * OpenAPI spec version: 0.1.0
 */
import type { Score } from "./score";

export interface RoundWithScores {
  id: number;
  courseId: number;
  courseName?: string;
  playerName: string;
  date: Date;
  totalScore?: number;
  totalPar?: number;
  scores: Score[];
  createdAt: Date;
}