# lib/api-zod/src/generated/types/score.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chameleon Golf API
 * OpenAPI spec version: 0.1.0
 */

export interface Score {
  id: number;
  roundId: number;
  holeNumber: number;
  par: number;
  strokes: number;
  notes?: string;
}